pyproject.toml
[tool.poetry]
name = "locate"
version = "0.1.0"
description = "A package for.."
authors = ["Lucrezia Valeriani"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
numpy = "<2"
matplotlib = "^3.9.2"
jupyter = "^1.0.0"
torch = {version = "^2.4.0", python = "^3.7", platform = "linux"}
numba = "^0.60.0"
scikit-learn = "^1.5.2"
pandas = "^2.2.3"
seaborn = "^0.13.2"
pyro-ppl = "^1.9.1"

[tool.poetry.dev-dependencies]

[tool.poetry.group.dev.dependencies]
myst-nb = {version = "^1.1.1", python = "^3.9"}
sphinx-autoapi = "^3.2.1"
sphinx-rtd-theme = "^2.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
